requirements.txt
Flask>=2.0.0
opencv-python>=4.5.5
numpy>=1.21.0
Pillow>=9.0.0
werkzeug>=2.2.0
gunicorn
